creating catalog file with book table sqlit database, it has 4 requests, get all books, get book by id, get books by topic, decreament books for the purchase requests
